Organizing WeeklyHoursAttendedCalculator
