+++ .rubocop_todo.yml
@@ -74,14 +74,3 @@ RSpec/VerifiedDoubles:
     - 'spec/lib/appsignal_reporting_spec.rb'
     - 'spec/services/wonderschool/necc/attendance_csv_importer_spec.rb'
     - 'spec/services/wonderschool/necc/dashboard_case_importer_spec.rb'
-
-# Offense count: 4
-# Cop supports --auto-correct.
-# Configuration parameters: Include.
-# Include: spec/**/*.rb, test/**/*.rb
-Rails/RedundantTravelBack:
-  Exclude:
-    - 'spec/blueprints/child_blueprint_spec.rb'
-    - 'spec/services/illinois_attendance_rate_calculator_spec.rb'
-    - 'spec/services/nebraska_absence_generator_spec.rb'
-    - 'spec/services/wonderschool/necc/onboarding_case_importer_spec.rb'
